+++ package.json
@@ -1,14 +1,19 @@
 {
-  "name": "investec-card-emu",
+  "name": "investec-card-cli",
   "version": "1.0.0",
   "main": "bin/index.js",
   "bin": {
     "card-emu": "./bin/index.js"
   },
+  "type": "module",
   "scripts": {
+    "build": "tsc",
     "test": "jest",
-    "prettier:check": "prettier --check .",
-    "prettier:fix": "prettier --write ."
+    "check-exports": "attw --pack . --ignore-rules=cjs-resolves-to-esm",
+    "check-format": "prettier --check .",
+    "ci": "npm run build && npm run check-format && npm run check-exports && npm run test",
+    "dev": "vitest",
+    "format": "prettier --write ."
   },
   "keywords": [
     "investec",
@@ -24,11 +29,15 @@
   "dependencies": {
     "chalk": "^4.1.2",
     "dotenv": "^16.3.1",
-    "programmable-card-code-emulator": "^1.2.0",
+    "programmable-card-code-emulator": "^1.4.2",
     "yargs": "^17.7.2"
   },
   "devDependencies": {
+    "@arethetypeswrong/cli": "^0.17.2",
     "jest": "^29.7.0",
     "prettier": "3.0.3"
-  }
+  },
+  "files": [
+    "dist"
+  ]
 }
